Fix logical error in AccountAgeProcessor
The code change fixes a logical error in the `execute` method of the `AccountAgeProcessor` class. The condition for checking the content type has been corrected from using an "or" operator to an "and" operator. This ensures that the condition is only satisfied when the content type is both not equal to `POST_TITLE` and not equal to `COMMENT`.